Build out a table and css grid version of the assessment overview
I think i'm going to use the css grid version since it's much better from a responsive design standpoint. To get the css-grid version working, I had to add vue-fragment (which I was considering for the table version as well). I should probably refactor this into something a bit more readable, but this was a big step in proving that it's possible