Migrate all routes to be under a /:studentId/ prefix
This way we can have permanent routes to a students info. Whether or not this is how we should proceed long term is still up in the air. The other way would be to remove this but create a `?studentId=ASDF` route param to do studentId switching. But that still prevents emailing links, which feels bad